Major changes to recursion and url detection
URLs are now detected in the returned JSON @id field, which is saved and
checked to make sure that we dont parse them again. We expect one
annotationList and one Canvas, may need to be changed.

Added a second text area just for annotations so that they can be
inserted separately from the canvas data

recursion no longer needs the multi object helper, because it is just an
extra step in recursion which is not needed.

Still need to parse only the label and the content for annotaitons
rather than every type.